requirements.txt
aiogram==3.13.1
google-generativeai==0.8.3
supabase==2.9.1
openpyxl==3.1.5
APScheduler==3.10.4
pydub==0.25.1
python-dotenv==1.0.1
pytz==2024.2
matplotlib==3.9.2
reportlab==4.2.5
cachetools==5.5.0
aiohttp==3.10.10
